+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.0.5] - 2016-10-20
+- Added `payment_status_request` feature support.
+- Removed schedule status check event, this will be part of the Pronamic iDEAL plugin.
+- Simplified payment methods constants code.
+- Added error status.
+
 ## [1.0.4] - 2016-06-08
 - Simplified the gateay payment start function.
 
@@ -26,7 +32,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2016-02-01
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.4...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.5...HEAD
+[1.0.5]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.4...1.0.5
 [1.0.4]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/wp-pay-gateways/ing-kassa-compleet/compare/1.0.1...1.0.2